requirements.txt
trimesh
pyrender
pyyaml
numpy
pycollada